Refactor collect.py and move public functions to collect_utils.py.
