+++ src/_pytask/collect_utils.py
@@ -1,23 +1,89 @@
-"""This module contains helpers for :mod:`_pytask.collect`."""
+"""This module provides utility functions for :mod:`_pytask.collect`."""
 from __future__ import annotations
 
 import itertools
 import uuid
 from pathlib import Path
 from typing import Any
+from typing import Callable
+from typing import Generator
 from typing import Iterable
+from typing import TYPE_CHECKING
 
 import attr
 from _pytask.exceptions import NodeNotCollectedError
+from _pytask.mark_utils import remove_marks
 from _pytask.nodes import find_duplicates
-from _pytask.nodes import MetaNode
-from _pytask.session import Session
+from pybaum.tree_util import tree_map
 
 
-__all__ = ["convert_objects_to_node_dictionary"]
+if TYPE_CHECKING:
+    from _pytask.session import Session
+    from _pytask.nodes import MetaNode
 
 
-def convert_objects_to_node_dictionary(objects: Any, when: str) -> dict[Any, Any]:
+__all__ = ["depends_on", "parse_nodes", "produces"]
+
+
+def depends_on(
+    objects: Any | Iterable[Any] | dict[Any, Any]
+) -> Any | Iterable[Any] | dict[Any, Any]:
+    """Specify dependencies for a task.
+
+    Parameters
+    ----------
+    objects : Union[Any, Iterable[Any], Dict[Any, Any]]
+        Can be any valid Python object or an iterable of any Python objects. To be
+        valid, it must be parsed by some hook implementation for the
+        :func:`pytask.hookspecs.pytask_collect_node` entry-point.
+
+    """
+    return objects
+
+
+def produces(
+    objects: Any | Iterable[Any] | dict[Any, Any]
+) -> Any | Iterable[Any] | dict[Any, Any]:
+    """Specify products of a task.
+
+    Parameters
+    ----------
+    objects : Union[Any, Iterable[Any], Dict[Any, Any]]
+        Can be any valid Python object or an iterable of any Python objects. To be
+        valid, it must be parsed by some hook implementation for the
+        :func:`pytask.hookspecs.pytask_collect_node` entry-point.
+
+    """
+    return objects
+
+
+def parse_nodes(
+    session: Session, path: Path, name: str, obj: Any, parser: Callable[..., Any]
+) -> Any:
+    objects = _extract_nodes_from_function_markers(obj, parser)
+    nodes = _convert_objects_to_node_dictionary(objects, parser.__name__)
+    nodes = tree_map(lambda x: _collect_node(session, path, name, x), nodes)
+    return nodes
+
+
+def _extract_nodes_from_function_markers(
+    function: Callable[..., Any], parser: Callable[..., Any]
+) -> Generator[Any, None, None]:
+    """Extract nodes from a marker.
+
+    The parser is a functions which is used to document the marker with the correct
+    signature. Using the function as a parser for the ``args`` and ``kwargs`` of the
+    marker provides the expected error message for misspecification.
+
+    """
+    marker_name = parser.__name__
+    _, markers = remove_marks(function, marker_name)
+    for marker in markers:
+        parsed = parser(*marker.args, **marker.kwargs)
+        yield parsed
+
+
+def _convert_objects_to_node_dictionary(objects: Any, when: str) -> dict[Any, Any]:
     """Convert objects to node dictionary."""
     list_of_dicts = [_convert_to_dict(x) for x in objects]
     _check_that_names_are_not_used_multiple_times(list_of_dicts, when)
@@ -49,45 +115,6 @@ def _convert_to_dict(x: Any, first_level: bool = True) -> Any | dict[Any, Any]:
         return {_Placeholder(scalar=True): x}
     else:
         return x
-
-
-def _collect_node(
-    session: Session, path: Path, name: str, node: str | Path
-) -> dict[str, MetaNode]:
-    """Collect nodes for a task.
-
-    Parameters
-    ----------
-    session : _pytask.session.Session
-        The session.
-    path : Path
-        The path to the task whose nodes are collected.
-    name : str
-        The name of the task.
-    nodes : Dict[str, Union[str, Path]]
-        A dictionary of nodes parsed from the ``depends_on`` or ``produces`` markers.
-
-    Returns
-    -------
-    Dict[str, MetaNode]
-        A dictionary of node names and their paths.
-
-    Raises
-    ------
-    NodeNotCollectedError
-        If the node could not collected.
-
-    """
-    collected_node = session.hook.pytask_collect_node(
-        session=session, path=path, node=node
-    )
-    if collected_node is None:
-        raise NodeNotCollectedError(
-            f"{node!r} cannot be parsed as a dependency or product for task "
-            f"{name!r} in {path!r}."
-        )
-
-    return collected_node
 
 
 def _check_that_names_are_not_used_multiple_times(
@@ -168,3 +195,42 @@ def _merge_dictionaries(list_of_dicts: list[dict[Any, Any]]) -> dict[Any, Any]:
                 out[k] = v
 
     return out
+
+
+def _collect_node(
+    session: Session, path: Path, name: str, node: str | Path
+) -> dict[str, MetaNode]:
+    """Collect nodes for a task.
+
+    Parameters
+    ----------
+    session : _pytask.session.Session
+        The session.
+    path : Path
+        The path to the task whose nodes are collected.
+    name : str
+        The name of the task.
+    nodes : Dict[str, Union[str, Path]]
+        A dictionary of nodes parsed from the ``depends_on`` or ``produces`` markers.
+
+    Returns
+    -------
+    Dict[str, MetaNode]
+        A dictionary of node names and their paths.
+
+    Raises
+    ------
+    NodeNotCollectedError
+        If the node could not collected.
+
+    """
+    collected_node = session.hook.pytask_collect_node(
+        session=session, path=path, node=node
+    )
+    if collected_node is None:
+        raise NodeNotCollectedError(
+            f"{node!r} cannot be parsed as a dependency or product for task "
+            f"{name!r} in {path!r}."
+        )
+
+    return collected_node
